+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 * Fixed an issue where the **format** command has incorrectly recognized on which files to run when running using git.
+* Added a validation to ensure reputation command's default argument is set as an array input.
 
 # 1.4.6
 * Fixed an issue where **validate** suggests, with no reason, running **format** on missing mandatory keys in yml file.